pyproject.toml
[project]
name = "clothing-brand-marketing-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.7.2",
    "crewai-tools>=1.7.2",
    "python-dotenv>=1.1.1",
]
